Adds pg_get_tablespace_ddl function
This patch provides a "retail" function for getting the DDL of any named
tablespace.

Currently, there exist ways of "wholesale" dumping the DDL for an entire
database or even cluster; this function will ideally be part of a suite
of similar "retail" functions for dumping the DDL of various database
objects.